# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Banque API",
        "description": "Documentation de l'API REST du projet Banque",
        "contact": {
            "name": "Abdoulaye Diome",
            "email": "abdoulaye.diome@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://bankt-1.onrender.com/api/v1",
            "description": "Serveur de production Render"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "description": "Authentifie un utilisateur (Admin ou Client) et retourne les tokens dans les cookies",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Cookies d'authentification (access_token et refresh_token)",
                                "schema": {
                                    "type": "string",
                                    "example": "access_token=eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9...; HttpOnly; Secure; SameSite=Lax, refresh_token=eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9...; HttpOnly; Secure; SameSite=Lax"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Rafraîchir le token d'accès",
                "description": "Utilise le refresh token pour obtenir un nouveau token d'accès",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Token rafraîchi avec succès",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Nouveaux cookies d'authentification (access_token et refresh_token)",
                                "schema": {
                                    "type": "string",
                                    "example": "access_token=eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9...; HttpOnly; Secure; SameSite=Lax, refresh_token=eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9...; HttpOnly; Secure; SameSite=Lax"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token rafraîchi avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "traceId": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion utilisateur",
                "description": "Déconnecte l'utilisateur et invalide tous ses tokens",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Cookies d'authentification supprimés (access_token et refresh_token)",
                                "schema": {
                                    "type": "string",
                                    "example": "access_token=; expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly; Secure; SameSite=Lax, refresh_token=; expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly; Secure; SameSite=Lax"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "traceId": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error Response",
                "description": "Format de réponse d'erreur standard",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue"
                    },
                    "errorCode": {
                        "type": "string",
                        "example": "VALIDATION_ERROR"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "path": {
                        "type": "string"
                    },
                    "traceId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "LoginRequest": {
                "title": "Login Request",
                "description": "Données de connexion",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Mot de passe",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "title": "Login Response",
                "description": "Réponse de connexion",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Connexion réussie"
                    },
                    "data": {
                        "properties": {
                            "user": {
                                "properties": {
                                    "id": {
                                        "description": "ID de l'utilisateur",
                                        "type": "string"
                                    },
                                    "nom": {
                                        "description": "Nom (pour Admin)",
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "description": "Prénom (pour Admin)",
                                        "type": "string"
                                    },
                                    "titulaire": {
                                        "description": "Nom complet (pour Client)",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Rôle de l'utilisateur",
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "client"
                                        ]
                                    }
                                },
                                "type": "object"
                            },
                            "token_type": {
                                "type": "string",
                                "example": "Bearer"
                            }
                        },
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "path": {
                        "type": "string"
                    },
                    "traceId": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification JWT - Utilisez le format: Bearer {token} ou les cookies d'authentification",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieAuth": {
                "type": "apiKey",
                "description": "Authentification via cookie - Le token est automatiquement inclus dans les requêtes",
                "name": "access_token",
                "in": "cookie"
            }
        }
    },
...
}